remove route_vans migration
