+++ backend/alembic/versions/311f80f8b8dd_create_routes_table.py
@@ -1,7 +1,7 @@
 """create routes table
 
 Revision ID: 311f80f8b8dd
-Revises: f1a02e4febad
+Revises: 174e5a759f2d
 Create Date: 2023-10-04 17:32:02.895752
 
 """
@@ -11,7 +11,7 @@ from alembic import op
 
 # revision identifiers, used by Alembic.
 revision: str = "311f80f8b8dd"
-down_revision: Union[str, None] = "f1a02e4febad"
+down_revision: Union[str, None] = "174e5a759f2d"
 branch_labels: Union[str, Sequence[str], None] = None
 depends_on: Union[str, Sequence[str], None] = None
