+++ backend/alembic/versions/f1a02e4febad_create_vans_table.py
@@ -1,7 +1,7 @@
 """create vans table
 
 Revision ID: f1a02e4febad
-Revises: 174e5a759f2d
+Revises: 311f80f8b8dd
 Create Date: 2023-10-04 17:31:57.484449
 
 """
@@ -11,7 +11,7 @@ from alembic import op
 
 # revision identifiers, used by Alembic.
 revision: str = "f1a02e4febad"
-down_revision: Union[str, None] = "174e5a759f2d"
+down_revision: Union[str, None] = "311f80f8b8dd"
 branch_labels: Union[str, Sequence[str], None] = None
 depends_on: Union[str, Sequence[str], None] = None
 
@@ -21,7 +21,11 @@ def upgrade() -> None:
         """
         CREATE TABLE IF NOT EXISTS public.vans (
 	        id serial PRIMARY KEY,
-	        wheelchair bool NOT NULL
+	        wheelchair bool NOT NULL,
+	        route_id int,
+	        FOREIGN KEY (route_id)
+		        REFERENCES routes (id)
+		        ON DELETE SET NULL
         );
     """
     )
